Add ability to view X amount of DB objects, and Add choosing show by ID
The GetShows method of DbManager now can be specified to return X number of shows, or by default will return all shows in the Db. It is also sorted by Descending order so that the newest items will appear on top. In UserInput there is also some new methods that allow for the user to choose a show from the DB by Id. This will enable us to both delete objects and edit objects.